aider: Added data folders for Grafana and Prometheus for persistent storage.
# Aider chat conversation:

USER: add data folder for grafana and prometheus

ASSISTANT: To add data folders for Grafana and Prometheus, I will update the `docker-compose.yml` file to include volume bindings for persistent storage. This will ensure that data for both services is stored on the host machine and persists across container restarts.

Here are the diffs for the changes:

```diff
--- docker-compose.yml
+++ docker-compose.yml
@@ ... @@
   prometheus:
     image: prom/prometheus:v2.33.1
     volumes:
       - ./config/prometheus.yml:/etc/prometheus/prometheus.yml
+      - ./data/prometheus:/prometheus
     ports:
       - "9090:9090"
     command:
       - '--config.file=/etc/prometheus/prometheus.yml'
       - '--storage.tsdb.path=/prometheus'
       - '--web.console.libraries=/usr/share/prometheus/console_libraries'
       - '--web.console.templates=/usr/share/prometheus/consoles'

   grafana:
     image: grafana/grafana:9.1.7
     volumes:
       - ./config/grafana.ini:/etc/grafana/grafana.ini
+      - ./data/grafana:/var/lib/grafana
     ports:
       - "3000:3000"
     depends_on:
       - prometheus
```

These changes will create two new directories, `./data/grafana` and `./data/prometheus`, which will be used to store Grafana and Prometheus data, respectively. Ensure that these directories are created on the host machine and have the correct permissions set for the Docker daemon to access them.